[refactor] : 폴더명 및 구조 정리
