+++ src/component/FoodForm.js
@@ -2,8 +2,8 @@ import { SELECT_OPTIONS } from "../constants/constants.js";
 import { CAPTION } from "../constants/systemMessage.js";
 import { foodItems } from "../mock/mockItem.js";
 import { FoodListPage } from "../pages/FoodListPage.js";
-import { removeError, resetError } from "../util/handleIsError.js";
-import { modalClose } from "../util/modalAction.js";
+import { removeError } from "../util/errorHandler.js";
+import { modalClose } from "../util/modalHandler.js";
 import { validateFoodItem } from "../validate/validateFoodItem.js";
 import { Button } from "./button/Button.js";
 import { ButtonContainer } from "./button/ButtonContainer.js";
