+++ templates/index.html
@@ -242,4 +242,7 @@
       </div>
     </main>
   </body>
+  <label for="link text-caption">${label}</label>
+  <input type="text" name="${name}" id="${name}" />
+  ${renderCaption(caption)}
 </html>
